+++ src/main/java/com/hamdihawari/server/exception/ErrorDetails.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.exception;
+
+public class ErrorDetails {
+}
