--- src/main/java/com/hamdihawari/server/language/Language.java
@@ -1,47 +1,0 @@
-package com.hamdihawari.server.language;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "language")
-public class Language {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-
-    @Column(nullable = false, unique = true)
-    @jakarta.validation.constraints.NotBlank(message = "Code is mandatory")
-    private String code;
-
-    @Column(nullable = false)
-    @jakarta.validation.constraints.NotBlank(message = "Name is mandatory")
-    private String name;
-
-    // Getters and Setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
-
