--- src/main/java/com/hamdihawari/server/language/LanguageRepository.java
@@ -1,6 +1,0 @@
-package com.hamdihawari.server.language;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LanguageRepository extends JpaRepository<Language, Long> {
-}
