+++ src/main/java/com/hamdihawari/server/language/controller/LanguageController.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.language.controller;
+
+public class LanguageController {
+}
