+++ src/main/java/com/hamdihawari/server/language/entity/Language.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.language.entity;
+
+public class Language {
+}
