+++ src/main/java/com/hamdihawari/server/language/repository/LanguageRepository.java
@@ -1,0 +1,9 @@
+package com.hamdihawari.server.language.languageRepository;
+
+import com.hamdihawari.server.language.entity.Language;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface LanguageRepository extends JpaRepository<Language, Long> {
+}
