+++ src/main/java/com/hamdihawari/server/language/service/LanguageService.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.language.service;
+
+public class LanguageService {
+}
