--- src/main/java/com/hamdihawari/server/project/imageGroup/ImageGroupRepository.java
@@ -1,9 +1,0 @@
-package com.hamdihawari.server.project.imageGroup;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ImageGroupRepository extends JpaRepository<ImageGroup, Long> {
-    List<ImageGroup> findByProjectDetailsId(Long projectDetailsId);
-}
