--- src/main/java/com/hamdihawari/server/project/projectCard/ProjectCardRepository.java
@@ -1,11 +1,0 @@
-package com.hamdihawari.server.project.projectCard;
-
-import com.hamdihawari.server.language.Language;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface ProjectCardRepository extends JpaRepository<ProjectCard, Long> {
-    List<ProjectCard> findByLanguage(Language language);
-}
-
