--- src/main/java/com/hamdihawari/server/project/projectDetails/ProjectDetailRepository.java
@@ -1,6 +1,0 @@
-package com.hamdihawari.server.project.projectDetails;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProjectDetailRepository extends JpaRepository<ProjectDetails, Long> {
-}
