+++ src/main/java/com/hamdihawari/server/projectCard/controller/ProjectCardController.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.projectCard.controller;
+
+public class ProjectCardController {
+}
