+++ src/main/java/com/hamdihawari/server/projectCard/controller/ProjectCardTranslationController.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.projectCard.controller;
+
+public class ProjectCardTranslationController {
+}
