+++ src/main/java/com/hamdihawari/server/projectCard/dto/ProjectCardTranslationDTO.java
@@ -1,0 +1,50 @@
+package com.hamdihawari.server.dto;
+
+public class ProjectCardTranslationDTO {
+    private Long id;
+    private Long projectCardId; // Foreign Key
+    private String language; // e.g., "en", "de", "ar"
+    private String title;
+    private String description;
+
+    // Getters and Setters
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Long getProjectCardId() {
+        return projectCardId;
+    }
+
+    public void setProjectCardId(Long projectCardId) {
+        this.projectCardId = projectCardId;
+    }
+
+    public String getLanguage() {
+        return language;
+    }
+
+    public void setLanguage(String language) {
+        this.language = language;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}
