+++ src/main/java/com/hamdihawari/server/projectCard/entity/ProjectCard.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.projectCard.entity;
+
+public class ProjectCard {
+}
