+++ src/main/java/com/hamdihawari/server/projectCard/repository/ProjectCardRepository.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.projectCard.repository;
+
+public interface ProjectCardRepository {
+}
