+++ src/main/java/com/hamdihawari/server/projectCard/service/ProjectCardService.java
@@ -1,0 +1,4 @@
+package com.hamdihawari.server.projectCard.service;
+
+public class ProjectCardService {
+}
